Created a new component for select pdf modal
