+++ superset-frontend/src/explore/components/useExploreAdditionalActionsMenu/index.jsx
@@ -48,9 +48,9 @@ import {
 import ViewQueryModal from '../controls/ViewQueryModal';
 import EmbedCodeContent from '../EmbedCodeContent';
 import DashboardsSubMenu from './DashboardsSubMenu';
+import PdfTemplateModal from './PdfTemplateModal.jsx';
 import { ChartClient } from '@superset-ui/core';
 import { useHistory } from 'react-router-dom';
-import ReactDOM from 'react-dom';
 import { makeApi, getClientErrorObject } from '@superset-ui/core';
 
 const MENU_KEYS = {
@@ -119,43 +119,6 @@ const iconReset = css`
     vertical-align: 0;
   }
 `;
-
-// Modal component for selecting a PDF template
-const PdfTemplateModal = ({ isOpen, onClose, onConfirm, templates, selectedTemplate, setSelectedTemplate }) => {
-  if (!isOpen) return null;
-
-  return ReactDOM.createPortal(
-    <div className="modal-overlay" style={{ position: 'fixed', top: 0, left: 0, right: 0, bottom: 0, backgroundColor: 'rgba(0,0,0,0.5)', display: 'flex', alignItems: 'center', justifyContent: 'center', zIndex: 1000 }}>
-      <div className="modal-content" style={{ backgroundColor: 'white', padding: '20px', borderRadius: '8px', width: '400px', maxWidth: '90%' }}>
-        <h2 style={{ marginBottom: '20px' }}>Select PDF Template</h2>
-        <div style={{ marginBottom: '20px' }}>
-          <label style={{ display: 'block', marginBottom: '8px' }}>Template:</label>
-          <select
-            value={selectedTemplate || ''}
-            onChange={(e) => setSelectedTemplate(e.target.value)}
-            style={{ width: '100%', padding: '8px', borderRadius: '4px', border: '1px solid #d9d9d9' }}
-          >
-            <option value="" disabled>Select a template</option>
-            {templates.map((template) => (
-              <option key={template.id} value={template.id}>
-                {template.name}
-              </option>
-            ))}
-          </select>
-        </div>
-        <div style={{ display: 'flex', justifyContent: 'flex-end', gap: '10px' }}>
-          <Button buttonStyle="secondary" onClick={onConfirm} disabled={!selectedTemplate}>
-            Confirm
-          </Button>
-          <Button buttonStyle="secondary" onClick={onClose}>
-            Cancel
-          </Button>
-        </div>
-      </div>
-    </div>,
-    document.body
-  );
-};
 
 export const useExploreAdditionalActionsMenu = (
   latestQueryFormData,
